requirements.txt
streamlit==1.24.0
pandas==2.0.3
sqlalchemy==2.0.18
openpyxl==3.1.2
pymysql==1.1.0
PyYAML==6.0
altair==5.0.1